requirements.txt
matplotlib>=3.1.3
numpy>=1.16.0
Pillow>=8.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
tensorboard==1.14.0
torch>=1.1.0
torchvision>=0.3.0
tqdm==4.42.1
